requirements.txt
pandas==1.3.4
Pillow==9.1.1
requests==2.26.0
streamlit==1.10.0
streamlit_option_menu==0.3.2
DateTime~=4.3
websocket~=0.2.1